% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutation.R
\name{apply_mutation}
\alias{apply_mutation}
\title{Processes the mutation object}
\usage{
apply_mutation(mutation, sequence)
}
\arguments{
\item{mutation}{string}

\item{sequence}{string}
}
\value{
An object with various info for mutation
}
\description{
This function is for the neoantigens project.
We call it with a mutation and sequence.
}
\examples{
apply_mutation("ins2AA","atagtagtagtagct")
}
\references{
\url{https://cran.r-project.org/web/packages/roxygen2/vignettes/roxygen2.html}
}
\author{
Rachel Alcraft, \url{https://github.com/instituteofcancerresearch/PkgStopGain}
}
