% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequences.R
\name{get_gene_seq}
\alias{get_gene_seq}
\title{Return the sequences given genes}
\usage{
get_gene_seq(gene)
}
\arguments{
\item{gene}{string}
}
\value{
A dna sequence as a string
}
\description{
This function is for the neoantigens project.
We call it with a gene.
}
\examples{
get_gene_seq("brca1")
}
\references{
\url{https://cran.r-project.org/web/packages/roxygen2/vignettes/roxygen2.html}
}
\author{
Rachel Alcraft, \url{https://github.com/instituteofcancerresearch/r-reversions}
}
