% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{get_pre_saved_genes_list}
\alias{get_pre_saved_genes_list}
\title{Returns the genes that have pre-saved dna}
\usage{
get_pre_saved_genes_list()
}
\value{
a list of genes with pre-saved dna sequences
}
\description{
This function is for the neoantigens project.
We call it with a mutation, gene and sequence.
}
\examples{
get_pre_saved_genes_list()
}
\references{
\url{https://cran.r-project.org/web/packages/roxygen2/vignettes/roxygen2.html}
}
\author{
Rachel Alcraft, \url{https://github.com/instituteofcancerresearch/PkgStopGain}
}
