% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stopgain.R
\name{stop_gain_factors}
\alias{stop_gain_factors}
\title{Return the upstream and downstream stop gains for a mutation}
\usage{
stop_gain_factors(mutation, sequence)
}
\arguments{
\item{mutation}{string}

\item{sequence}{string}

\item{gene}{string}
}
\value{
An object with various info for the stop gain
}
\description{
This function is for the neoantigens project.
We call it with a mutation, gene and sequence.
}
\examples{
stop_gain_factors("ins2AA","atagtagtagtagct")
}
\references{
\url{https://cran.r-project.org/web/packages/roxygen2/vignettes/roxygen2.html}
}
\author{
Rachel Alcraft, \url{https://github.com/instituteofcancerresearch/PkgStopGain}
}
