% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updown.R
\name{up_down_factors}
\alias{up_down_factors}
\title{Return the upstream and downstream neoantigen information along with distance and sequence}
\usage{
up_down_factors(mutation, sequence)
}
\arguments{
\item{mutation}{string}

\item{sequence}{string}

\item{gene}{string}
}
\value{
An object with various info
}
\description{
This function is for the neoantigens project.
We call it with a mutation, gene and sequence.
}
\examples{
ff_rtest_add_basic(1, 1)
ff_rtest_add_basic(c(10,10), c(1,3))
}
\references{
\url{https://cran.r-project.org/web/packages/roxygen2/vignettes/roxygen2.html}
}
\seealso{
\code{\link[=ff_rtest_mul_basic]{ff_rtest_mul_basic()}} for multiplication.
}
\author{
Rachel Alcraft, \url{https://github.com/instituteofcancerresearch/r-reversions}
}
